Renaming files as per feedback
